bug fixed

+++ MyTask/model/operation.py
@@ -6,7 +6,7 @@ Created on Feb 4, 2013
 @author: GoldRatio
 '''
 from core.database import db
-from sqlalchemy import Column, Integer, String, DateTime, Boolean, Table, ForeignKey, Enum
+from sqlalchemy import Column, SmallInteger, BigInteger, String, DateTime, Boolean, Table, ForeignKey, Enum
 from sqlalchemy.orm import relationship
 
 __all__ = ['Operation']
@@ -16,17 +16,17 @@ class Operation(db.Model):
     operation_type_array = ['创建', '发起', '回复', '删除', '编辑', '查询', '回复', '开始', '暂停', '完成', '取消完成', '上传']
     target_type_array = ['项目', '讨论', '评论', '任务列表', '任务', '用户', '文件']
 
-    id = Column(Integer, primary_key=True)
-    own_id = Column(Integer, ForeignKey('user.id'))
+    id = Column(BigInteger, primary_key=True)
+    own_id = Column(BigInteger, ForeignKey('user.id'))
     createTime = Column(DateTime)
-    operation_type = Column(Integer(2))
-    target_type = Column(Integer(2))
-    target_id = Column(Integer)
+    operation_type = Column(SmallInteger)
+    target_type = Column(SmallInteger)
+    target_id = Column(BigInteger)
     title = Column(String(100))
     url = Column(String(100))
     digest = Column(String(200))
-    team_id = Column(Integer)
-    project_id = Column(Integer, ForeignKey('project.id'))
+    team_id = Column(BigInteger)
+    project_id = Column(BigInteger, ForeignKey('project.id'))
 
     @property
     def operation_type_name(self):
